+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Do not log images exceptions as fatal errors [#722](https://github.com/rokwire/app-flutter-plugin/issues/722).
 - Fixed crash in SwipeDetector widget [#725](https://github.com/rokwire/app-flutter-plugin/issues/725).
 - WEB: Display the proper image after url is changed [#6032](https://github.com/rokwire/illinois-app/issues/6032).
+### Changed
+- Rework refresh token processing [#5836](https://github.com/rokwire/illinois-app/issues/5836).
 
 ## [1.13.5] - 2026-08-11
 ### Added